Tagged templating is all the rage, so why not
This adds in support for building a VTree and passing it into diffHTML.
It's allowed in any existing method like outerHTML and innerHTML. You
can use it like so:

    import { outerHTML, html } from 'diffhtml';

    function render(date) {
      outerHTML(document.body, html`<body>
        <h1>Some new content diff/patched in!</h1>

        <p>
          <span>Current date: ${date}</span>
        </p>
      </body>`);
    }

    setInterval(function() {
      render(new Date().toLocaleString());
    }, 1000);

    render(new Date().toLocaleString());

I'll start working on a Babel transform as well to make the entire parse
time disappear. Performance-ahoy. Thx JSX, yo-yo, bel for inspiration
choo-choo.